Add logger to the file watcher constructor
Having a default slf4j logger for `DirectoryWatcher` forces downstream
projects to use global loggers and provide a configuration for logback
or log4j implementors.

This pull request removes the default per-class logger and adds the
logger instance to the constructor `create` of `DirectoryWatcher` so
that users can control how file watching events are logged.

This change makes for a better, more explicit API.